requirements.txt
tensorflow==2.13.0
Pillow==9.1.1
numpy==1.22.4
streamlit~=1.30.0
keras~=2.13.1
PyYAML~=6.0.1
pydantic~=2.5.3
s3fs~=2023.12.2
kaggle~=1.6.2
h5py~=3.10.0
boto3~=1.34.18